R Package: roundDecimal - Fast Rounding of Decimal Points
Created as an extension to the Round series of functions as functions ceiling, floor, and trunc do not currently allow the user to specify the number of decimal places to apply the functions to.  roundDecimal provides the user with a simple strategy for rounding a single or array of decimal(s) to a specified number of decimal places.  The user is able to choose from one of three methods of rounding including: normal rounding, rounding up, rounding down, or truncation.