R Package: roundDecimal - Fast Rounding of Decimal Points (Version 2)
Created as an extension to the Round series of functions as functions ceiling, floor, and trunc do not currently allow the user to specify the number of decimal places to apply the functions to.  roundDecimal provides the user with a simple strategy for rounding a single or array of decimal(s) to a specified number of decimal places.  The user is able to choose from one of three methods of rounding including: normal rounding, rounding up, rounding down, or truncation.

This version of roundDecimal has a median run time which is approximately 850 times faster than previous versions (1.0 and 1.1).

I ran a computational speed test using microbenchmark.  The data used contained a vector of 1,000,000 random numbers with decimal points. The median computation speeds reported from 1,000 runs of each function (versions 1.1 and 2.0) were 4,905.03625 ms and 5.78085 ms respectively. Under these test conditions, roundDecimal version 2.0 had a median run time which was 848.4974 times faster than version 1.1.

I ensured both functions produce identical output using the identical() function on the two output arrays.